fix/clean-examples
