+++ pipelex.toml
@@ -1,3 +1,9 @@
+[pipelex.feature_config]
+# WIP/Experimental feature flags
+is_pipeline_tracking_enabled = true
+is_activity_tracking_enabled = true
+is_reporting_enabled = true
+
 [pipelex.aws_config]
 api_key_method = "env" 
 # The possible values are "env" and "secret_provider". 
@@ -49,3 +55,4 @@ api_key_method = "env"
 [cogt.ocr_config]
 ocr_handles = ["mistral/mistral-ocr-latest"]
 page_output_text_file_name = "page_text.md"
+
