+++ pipelex_libraries/pipelines/examples/invoice_extractor/invoice.toml
@@ -1,0 +1,55 @@
+domain = "invoice_extraction"
+
+[concept]
+Invoice = "Invoice information extracted from text"
+InvoiceDetails = "The category of the invoice"
+
+[pipe]
+[pipe.process_invoice]
+PipeSequence = "Process relevant information from an invoice"
+inputs = { invoice_pdf = "native.PDF" }
+output = "Invoice"
+steps = [
+    { pipe = "extract_text_from_image", result = "invoice_pages" },
+    { pipe = "extract_invoice", batch_over = "invoice_pages", batch_as = "invoice_page", result = "invoice" },
+]
+
+[pipe.extract_text_from_image]
+PipeOcr = "Extract page contents from an image"
+inputs = { invoice_pdf = "native.PDF" }
+output = "native.Page"
+page_views = true
+
+[pipe.extract_invoice]
+PipeSequence = "Extract invoice information from an invoice text transcript"
+inputs = { invoice_page = "native.Page" }
+output = "Invoice"
+steps = [
+    { pipe = "analyze_invoice", result = "invoice_details" },
+    { pipe = "extract_invoice_data", result = "invoice" }
+]
+
+[pipe.analyze_invoice]
+PipeLLM = "Analyze the invoice"
+inputs = { invoice_page = "native.Page" }
+output = "InvoiceDetails"
+images = ["invoice_page.page_view"]
+prompt_template = """
+Analyze this invoice:
+
+@invoice_page.text_and_images.text.text
+"""
+
+[pipe.extract_invoice_data]
+PipeLLM = "Extract invoice information from an invoice text transcript"
+inputs = { invoice_page = "native.Page", invoice_details = "InvoiceDetails" }
+output = "Invoice"
+llm = "llm_to_extract_invoice"
+images = ["invoice_page.page_view"]
+prompt_template = """
+Extract invoice information from this invoice:
+
+The category of this invoice is: $invoice_details.category.
+
+@invoice_page.text_and_images.text.text
+"""
